% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlstrOpalr-package.R
\docType{package}
\name{mlstrOpalr-package}
\alias{mlstrOpalr}
\alias{mlstrOpalr-package}
\title{mlstrOpalr: Support compatibility between Maelstrom R packages and Opal environment}
\description{
The goal of mlstrOpalr is to provide functions to support compatibility between Maelstrom R packages and Opal environment. Opal is the OBiBa application for data management, and Opal environments can be integrated with RStudio environments using the package `opalr`. The user must have adequate credentials to interact with a specific Opal environment. Please see \url{https://opaldoc.obiba.org/en/dev/} for complete documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maelstrom-research/mlstrOpalr/}
  \item Report bugs at \url{https://github.com/maelstrom-research/mlstrOpalr/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Fabre \email{guijoseph.fabre@gmail.com} (\href{https://orcid.org/0000-0002-0124-9970}{ORCID})

Other contributors:
\itemize{
  \item Maelstrom-research group \email{info@maelstrom-research.org} [copyright holder, funder]
  \item OBiBa group \email{info@obiba.org} [contributor]
}

}
\keyword{internal}
