% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R
\name{mlstrOpalr_website}
\alias{mlstrOpalr_website}
\title{Call the help center for full documentation}
\usage{
mlstrOpalr_website()
}
\value{
Nothing to be returned. The function opens a web page.
}
\description{
This function is a direct call to the documentation in the repository hosting
the package. The user can access the description of the latest version of the
package, the vignettes, and the list of functions.
}
\examples{
{

mlstrOpalr_website()

}

}
