% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-import-from-packages.R
\docType{import}
\name{opal.login}
\alias{opal.login}
\title{Objects exported from other packages}
\keyword{exported}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{opalr}{\code{\link[opalr]{opal.login}}}
}}

